+++ 8382/Yankin Daniil/TheGame/Landscape/Forest.h
@@ -8,5 +8,5 @@ class Forest : public Landscape
 public:
     Forest(Point point);
     bool movementAccess() override;
-    uint8_t getObjectType() override;
+    uint16_t getObjectType() override;
 };
